Updating Libft and Turning on -O1 (#20)
* Updating Libft and Turning on -O1

Libft header renames (for shoter header names)
Libft preformance update
Turing on -O1 for optimsation

* made compatible with new master

* optimised some SDL functions to use memset4

+++ libft/src/color/ft_colorlcode.c
@@ -11,10 +11,10 @@
 /* ************************************************************************** */
 
 #include "ft_color.h"
-#include "ft_string.h"
-#include "ft_stringarray.h"
+#include "ft_str.h"
+#include "ft_strarr.h"
 
-ssize_t		ft_colorlcode(const char *str)
+int		ft_colorlcode(const char *str)
 {
 	char	**colors;
 	size_t	i;
